move ugly impl; cargo fmt
