+++ crates/motoko/src/lib/vm_types.rs
@@ -465,219 +465,6 @@ pub trait ActiveBorrow {
     }
 }
 
-impl Active for Core {
-    fn schedule_choice<'a>(&'a self) -> &'a ScheduleChoice {
-        &self.schedule_choice
-    }
-    fn cont<'a>(&'a mut self) -> &'a mut Cont {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.active.cont,
-            Actor(ref n) => {
-                &mut self
-                    .actors
-                    .map
-                    .get_mut(n)
-                    .unwrap()
-                    .active
-                    .as_mut()
-                    .unwrap()
-                    .cont
-            }
-        }
-    }
-    fn cont_source<'a>(&'a mut self) -> &'a mut Source {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.active.cont_source,
-            Actor(ref n) => {
-                &mut self
-                    .actors
-                    .map
-                    .get_mut(n)
-                    .unwrap()
-                    .active
-                    .as_mut()
-                    .unwrap()
-                    .cont_source
-            }
-        }
-    }
-    fn cont_prim_type<'a>(&'a mut self) -> &'a mut Option<PrimType> {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.active.cont_prim_type,
-            Actor(ref n) => {
-                &mut self
-                    .actors
-                    .map
-                    .get_mut(n)
-                    .unwrap()
-                    .active
-                    .as_mut()
-                    .unwrap()
-                    .cont_prim_type
-            }
-        }
-    }
-    fn env<'a>(&'a mut self) -> &'a mut Env {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.active.env,
-            Actor(ref n) => {
-                &mut self
-                    .actors
-                    .map
-                    .get_mut(n)
-                    .unwrap()
-                    .active
-                    .as_mut()
-                    .unwrap()
-                    .env
-            }
-        }
-    }
-    fn stack<'a>(&'a mut self) -> &'a mut Stack {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.active.stack,
-            Actor(ref n) => {
-                &mut self
-                    .actors
-                    .map
-                    .get_mut(n)
-                    .unwrap()
-                    .active
-                    .as_mut()
-                    .unwrap()
-                    .stack
-            }
-        }
-    }
-    fn store<'a>(&'a mut self) -> &'a mut Store {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.store,
-            Actor(ref n) => &mut self.actors.map.get_mut(n).unwrap().store,
-        }
-    }
-    fn debug_print_out<'a>(&'a mut self) -> &'a mut Vector<DebugPrintLine> {
-        &mut self.debug_print_out
-    }
-    fn counts<'a>(&'a mut self) -> &'a mut Counts {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &mut self.agent.counts,
-            Actor(ref n) => &mut self.actors.map.get_mut(n).unwrap().counts,
-        }
-    }
-}
-
-impl ActiveBorrow for Core {
-    fn cont<'a>(&'a self) -> &'a Cont {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => &self.agent.active.cont,
-            Actor(ref n) => {
-                & self
-                    .actors
-                    .map
-                    .get(n)
-                    .unwrap()
-                    .active
-                    .as_ref()
-                    .unwrap()
-                    .cont
-            }
-        }
-    }
-    fn cont_source<'a>(&'a self) -> &'a Source {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => & self.agent.active.cont_source,
-            Actor(ref n) => {
-                & self
-                    .actors
-                    .map
-                    .get(n)
-                    .unwrap()
-                    .active
-                    .as_ref()
-                    .unwrap()
-                    .cont_source
-            }
-        }
-    }
-    fn cont_prim_type<'a>(&'a self) -> &'a Option<PrimType> {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => & self.agent.active.cont_prim_type,
-            Actor(ref n) => {
-                & self
-                    .actors
-                    .map
-                    .get(n)
-                    .unwrap()
-                    .active
-                    .as_ref()
-                    .unwrap()
-                    .cont_prim_type
-            }
-        }
-    }
-    fn env<'a>(&'a self) -> &'a Env {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => & self.agent.active.env,
-            Actor(ref n) => {
-                & self
-                    .actors
-                    .map
-                    .get(n)
-                    .unwrap()
-                    .active
-                    .as_ref()
-                    .unwrap()
-                    .env
-            }
-        }
-    }
-    fn stack<'a>(&'a self) -> &'a Stack {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => & self.agent.active.stack,
-            Actor(ref n) => {
-                & self
-                    .actors
-                    .map
-                    .get(n)
-                    .unwrap()
-                    .active
-                    .as_ref()
-                    .unwrap()
-                    .stack
-            }
-        }
-    }
-    fn store<'a>(&'a self) -> &'a Store {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => & self.agent.store,
-            Actor(ref n) => & self.actors.map.get(n).unwrap().store,
-        }
-    }
-    fn debug_print_out<'a>(&'a self) -> &'a Vector<DebugPrintLine> {
-        & self.debug_print_out
-    }
-    fn counts<'a>(&'a self) -> &'a Counts {
-        use ScheduleChoice::*;
-        match &self.schedule_choice {
-            Agent => & self.agent.counts,
-            Actor(ref n) => & self.actors.map.get(n).unwrap().counts,
-        }
-    }
-}
-
 // Some ideas of how we could count and limit what the VM does,
 // to interject some "slow interactivity" into its execution.
 #[derive(Clone, Debug, Serialize, Deserialize)]
