+++ crates/motoko_proc_macro/tests/test_eval_open_block.rs
@@ -10,15 +10,14 @@ use test_log::test; // enable logging output for tests by default.
 fn test_eval_open_block() {
     let mut core = Core::new(parse_static!("let x = 666; let y = 777;").clone());
     core.continue_(&Limits::none()).unwrap();
-    core
-        .eval_open_block(
-            vec![
-                ("x", Value::Nat(BigUint::from(1_u32)).share()),
-                ("y", Value::Nat(BigUint::from(2_u32)).share()),
-            ],
-            parse_static!("var z = x + y").clone(),
-        )
-        .unwrap();
+    core.eval_open_block(
+        vec![
+            ("x", Value::Nat(BigUint::from(1_u32)).share()),
+            ("y", Value::Nat(BigUint::from(2_u32)).share()),
+        ],
+        parse_static!("var z = x + y").clone(),
+    )
+    .unwrap();
     let r = core.eval_prog(parse_static!("x + y").clone()).unwrap();
     assert_eq!(r, Value::Nat(BigUint::from(666 + 777_u32)).share());
 }
@@ -38,11 +37,10 @@ fn test_hashmap_performance_steps() {
 
     // generate initial data / batch random put.
     let size = 10;
-    core
-        .eval_open_block(
-            vec![("size", Value::Nat(BigUint::from(size as u32)).share())],
-            parse_static!(
-                "
+    core.eval_open_block(
+        vec![("size", Value::Nat(BigUint::from(size as u32)).share())],
+        parse_static!(
+            "
       var i = prim \"fastRandIterNew\" (?size, 1);
       let j = {
         next = func () { prim \"fastRandIterNext\" i }
@@ -53,8 +51,8 @@ fn test_hashmap_performance_steps() {
         map := m;
       }
     "
-            )
-            .clone(),
         )
-        .unwrap();
+        .clone(),
+    )
+    .unwrap();
 }
